+++ lib/td_dq/rules/audit.ex
@@ -1,0 +1,30 @@
+defmodule TdDq.Rules.Audit do
+  @moduledoc """
+  The Rules Audit context. The public functions in this module are designed to
+  be called using `Ecto.Multi.run/5`, although the first argument (`repo`) is
+  not currently used.
+  """
+
+  import TdDq.Audit.AuditSupport, only: [publish: 4, publish: 5]
+
+  @doc """
+  Publishes a `:rule_created` event. Should be called using `Ecto.Multi.run/5`.
+  """
+  def rule_created(_repo, %{rule: %{id: id}}, %{} = changeset, user_id) do
+    publish("rule_created", "rule", id, user_id, changeset)
+  end
+
+  @doc """
+  Publishes a `:rule_updated` event. Should be called using `Ecto.Multi.run/5`.
+  """
+  def rule_updated(_repo, %{rule: %{id: id}}, %{} = changeset, user_id) do
+    publish("rule_updated", "rule", id, user_id, changeset)
+  end
+
+  @doc """
+  Publishes a `:rule_deleted` event. Should be called using `Ecto.Multi.run/5`.
+  """
+  def rule_deleted(_repo, %{rule: %{id: id}}, user_id) do
+    publish("rule_deleted", "rule", id, user_id)
+  end
+end